feat: applied user identification
Added so the function can recognize whose at the time logging in as a user

+++ db/repository.py
@@ -1,20 +1,21 @@
 from db.database import db_dependency
 from db.models import Users, Roles
 from handler import exception
-from handler.auth import ctx, authenticate_user, create_token
+from handler.auth import ctx, authenticate_user, create_token, check_admin_role
 from schema import Token
 
 from typing import Annotated
 from fastapi import Depends
+from sqlalchemy.orm import Session
 from fastapi.security import OAuth2PasswordRequestForm
 
-def get_user(username: str, db: db_dependency):
+def get_user(username: str, db: Session):
     user = db.query(Users).filter(Users.username == username).first()
     if not user:
         exception.user_not_found()
     return user
 
-def create_user(username: str, password: str, role_id: int, db: db_dependency):
+def create_user(username: str, password: str, role_id: int, db: Session):
     user = db.query(Users).filter(Users.username == username).first()
     if user:
         exception.account_is_already_exist()
@@ -29,7 +30,7 @@ def create_user(username: str, password: str, role_id: int, db: db_dependency):
     db.commit()
     return "Successfully create account!"
 
-def login_user(form_data: Annotated[OAuth2PasswordRequestForm, Depends()], db: db_dependency):
+def login_user(form_data: OAuth2PasswordRequestForm, db: Session):
     user = authenticate_user(form_data.username, form_data.password, db)
     if not user:
         exception.user_not_found()
@@ -40,14 +41,12 @@ def login_user(form_data: Annotated[OAuth2PasswordRequestForm, Depends()], db: d
         token_type = 'bearer'
     )
 
-def change_username(old_name: str, new_name: str, password: str, db: db_dependency):
-    user = authenticate_user(old_name, password, db)
+def change_username(user: Users, new_name: str, db: Session):
     user.username = new_name
     db.commit()
     return f"Username is successfully changed"
 
-def change_password(username: str, old_pass: str, new_pass: str, db: db_dependency):
-    user = get_user(username, db)
+def change_password(user: Users, old_pass: str, new_pass: str, db: Session):
     if not user:
         exception.user_not_found()
     if not ctx.verify(old_pass, user.password):
@@ -56,13 +55,14 @@ def change_password(username: str, old_pass: str, new_pass: str, db: db_dependen
     db.commit()
     return f"Password is successfully changed"
 
-def get_all_user(db: db_dependency):
+def get_all_user(admin: Users, db: Session):
+    check_admin_role(admin)
     result = db.query(Users, Roles.role_name).join(Roles).filter(Users.role_id == Roles.role_id).all()
     return [{"user_id": user.id, "username": user.username, "password": user.password, "role_name": role_name} for user, role_name in result]
 
-def delete_user(username: str, db: db_dependency):
+def delete_user(admin: Users, username: str, db: Session):
+    check_admin_role(admin)
     user = get_user(username, db)
     db.delete(user)
     db.commit()
     return f"User with username '{username}' has been deleted successfully"
-
